aes: 2018 edition and `block-cipher` crate upgrade
Upgrades the `aes`, `aes-soft`, and `aesni` crates to Rust 2018 edition.

Additionally upgrades them to the (now 2018 edition) `block-cipher`
crate (formerly `block-cipher-trait`).
