+++ Cargo.toml
@@ -14,3 +14,7 @@ members = [
     "twofish",
     "threefish",
 ]
+
+[patch.crates-io]
+block-cipher = { git = "https://github.com/RustCrypto/traits.git" }
+stream-cipher = { git = "https://github.com/RustCrypto/traits.git" }
