+++ aes/aesni/Cargo.toml
@@ -1,22 +1,23 @@
 [package]
 name = "aesni"
 version = "0.6.0"
+description = "AES (Rijndael) block ciphers implementation using AES-NI"
 authors = ["RustCrypto Developers"]
 license = "MIT OR Apache-2.0"
-description = "AES (Rijndael) block ciphers implementation using AES-NI"
+edition = "2018"
 documentation = "https://docs.rs/aesni"
 repository = "https://github.com/RustCrypto/block-ciphers"
 keywords = ["crypto", "aes", "rijndael", "block-cipher"]
 categories = ["cryptography", "no-std"]
 
 [dependencies]
-block-cipher-trait = "0.6"
+block-cipher = "= 0.7.0-pre"
 opaque-debug = "0.2"
-stream-cipher = { version = "0.3", optional = true }
+stream-cipher = { version = "0.4.0-pre", optional = true }
 
 [dev-dependencies]
-block-cipher-trait = { version = "0.6", features = ["dev"] }
-stream-cipher = { version = "0.3", features = ["dev"] }
+block-cipher = { version = "0.7.0-pre", features = ["dev"] }
+stream-cipher = { version = "= 0.4.0-pre", features = ["dev"] }
 
 [features]
 default = ["ctr"]
