+++ aes/aesni/benches/aes192.rs
@@ -1,10 +1,10 @@
 #![no_std]
 #![cfg(any(target_arch = "x86_64", target_arch = "x86"))]
 #![feature(test)]
-extern crate aesni;
+
 extern crate test;
 
-use aesni::block_cipher_trait::BlockCipher;
+use aesni::block_cipher::{BlockCipher, NewBlockCipher};
 use aesni::Aes192;
 
 #[bench]
