+++ aes/aesni/src/aes128.rs
@@ -1,23 +1,23 @@
-use arch::*;
-use block_cipher_trait::generic_array::typenum::{U16, U24, U8};
-use block_cipher_trait::generic_array::GenericArray;
-use block_cipher_trait::BlockCipher;
+use crate::arch::*;
+use block_cipher::generic_array::typenum::{U16, U8};
+use block_cipher::generic_array::GenericArray;
+use block_cipher::{BlockCipher, NewBlockCipher};
 
+use crate::utils::{Block128, Block128x8};
 use core::mem;
-use utils::{Block128, Block128x8};
 
 mod expand;
 #[cfg(test)]
 mod test_expand;
 
-/// AES-192 block cipher
+/// AES-128 block cipher
 #[derive(Copy, Clone)]
-pub struct Aes192 {
-    encrypt_keys: [__m128i; 13],
-    decrypt_keys: [__m128i; 13],
+pub struct Aes128 {
+    encrypt_keys: [__m128i; 11],
+    decrypt_keys: [__m128i; 11],
 }
 
-impl Aes192 {
+impl Aes128 {
     #[inline(always)]
     pub(crate) fn encrypt8(&self, mut blocks: [__m128i; 8]) -> [__m128i; 8] {
         let keys = self.encrypt_keys;
@@ -32,9 +32,7 @@ impl Aes192 {
             aesenc8!(blocks, keys[7]);
             aesenc8!(blocks, keys[8]);
             aesenc8!(blocks, keys[9]);
-            aesenc8!(blocks, keys[10]);
-            aesenc8!(blocks, keys[11]);
-            aesenclast8!(blocks, keys[12]);
+            aesenclast8!(blocks, keys[10]);
         }
         blocks
     }
@@ -53,20 +51,16 @@ impl Aes192 {
             block = _mm_aesenc_si128(block, keys[7]);
             block = _mm_aesenc_si128(block, keys[8]);
             block = _mm_aesenc_si128(block, keys[9]);
-            block = _mm_aesenc_si128(block, keys[10]);
-            block = _mm_aesenc_si128(block, keys[11]);
-            _mm_aesenclast_si128(block, keys[12])
+            _mm_aesenclast_si128(block, keys[10])
         }
     }
 }
 
-impl BlockCipher for Aes192 {
-    type KeySize = U24;
-    type BlockSize = U16;
-    type ParBlocks = U8;
+impl NewBlockCipher for Aes128 {
+    type KeySize = U16;
 
     #[inline]
-    fn new(key: &GenericArray<u8, U24>) -> Self {
+    fn new(key: &GenericArray<u8, U16>) -> Self {
         let key = unsafe { mem::transmute(key) };
         let (encrypt_keys, decrypt_keys) = expand::expand(key);
         Self {
@@ -74,6 +68,11 @@ impl BlockCipher for Aes192 {
             decrypt_keys,
         }
     }
+}
+
+impl BlockCipher for Aes128 {
+    type BlockSize = U16;
+    type ParBlocks = U8;
 
     #[inline]
     fn encrypt_block(&self, block: &mut Block128) {
@@ -89,9 +88,7 @@ impl BlockCipher for Aes192 {
         let keys = self.decrypt_keys;
         unsafe {
             let mut b = _mm_loadu_si128(block.as_ptr() as *const __m128i);
-            b = _mm_xor_si128(b, keys[12]);
-            b = _mm_aesdec_si128(b, keys[11]);
-            b = _mm_aesdec_si128(b, keys[10]);
+            b = _mm_xor_si128(b, keys[10]);
             b = _mm_aesdec_si128(b, keys[9]);
             b = _mm_aesdec_si128(b, keys[8]);
             b = _mm_aesdec_si128(b, keys[7]);
@@ -119,9 +116,7 @@ impl BlockCipher for Aes192 {
         let keys = self.decrypt_keys;
         unsafe {
             let mut b = load8!(blocks);
-            xor8!(b, keys[12]);
-            aesdec8!(b, keys[11]);
-            aesdec8!(b, keys[10]);
+            xor8!(b, keys[10]);
             aesdec8!(b, keys[9]);
             aesdec8!(b, keys[8]);
             aesdec8!(b, keys[7]);
@@ -137,4 +132,4 @@ impl BlockCipher for Aes192 {
     }
 }
 
-impl_opaque_debug!(Aes192);
+impl_opaque_debug!(Aes128);
