+++ aes/aesni/src/aes128/expand.rs
@@ -1,6 +1,6 @@
-use arch::*;
+use crate::arch::*;
 
-use core::mem;
+use core::mem::MaybeUninit;
 
 macro_rules! expand_round {
     ($enc_keys:expr, $dec_keys:expr, $pos:expr, $round:expr) => {
@@ -27,8 +27,8 @@ macro_rules! expand_round {
 #[inline(always)]
 pub(super) fn expand(key: &[u8; 16]) -> ([__m128i; 11], [__m128i; 11]) {
     unsafe {
-        let mut enc_keys: [__m128i; 11] = mem::uninitialized();
-        let mut dec_keys: [__m128i; 11] = mem::uninitialized();
+        let mut enc_keys: [__m128i; 11] = MaybeUninit::uninit().assume_init();
+        let mut dec_keys: [__m128i; 11] = MaybeUninit::uninit().assume_init();
 
         let k = _mm_loadu_si128(key.as_ptr() as *const __m128i);
         _mm_store_si128(enc_keys.as_mut_ptr(), k);
