+++ aes/aesni/src/utils.rs
@@ -1,10 +1,10 @@
 #[cfg(test)]
-use arch::__m128i;
+use crate::arch::__m128i;
 #[cfg(test)]
 use core::mem;
 
-use block_cipher_trait::generic_array::typenum::{U16, U8};
-use block_cipher_trait::generic_array::GenericArray;
+use block_cipher::generic_array::typenum::{U16, U8};
+use block_cipher::generic_array::GenericArray;
 
 pub type Block128 = GenericArray<u8, U16>;
 pub type Block128x8 = GenericArray<GenericArray<u8, U16>, U8>;
